DIRECTOR: Fix Window::invertChannel in 32-bit mode

+++ engines/director/window.cpp
@@ -116,7 +116,7 @@ void Window::invertChannel(Channel *channel, const Common::Rect &destRect) {
 
 		for (int i = 0; i < srcRect.height(); i++) {
 			uint32 *src = (uint32 *)_composeSurface->getBasePtr(srcRect.left, srcRect.top + i);
-			const uint32 *msk = mask ? (const uint32 *)mask->getBasePtr(xoff, yoff + i) : nullptr;
+			const byte *msk = mask ? (const byte *)mask->getBasePtr(xoff, yoff + i) : nullptr;
 
 			for (int j = 0; j < srcRect.width(); j++, src++)
 				if (!mask || (msk && (*msk++)))
